pyproject.toml
[project]
name = "tennis-autoencoder-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops",
    "hydra-core>=1.3.2",
    "huggingface_hub",
    "lightning>=2.6.1",
    "ftfy==6.1.1",
    "iopath>=0.1.10",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.13.0.92",
    "regex",
    "tensorboard>=2.20.0",
    "timm>=1.0.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "yt-dlp>=2026.3.17",
]
